+++ src/Search/Output/ExportSearchOutput.php
@@ -183,7 +183,7 @@ abstract class ExportSearchOutput extends AbstractSearchOutput
         return $out;
     }
 
-    public static function displayData(array $data, array $params = [])
+    public function displayData(array $data, array $params = [])
     {
         /** @var array $CFG_GLPI */
         global $CFG_GLPI;